+++ test/appium_driver_test.rb
@@ -24,7 +24,7 @@ class AppiumDriverTest < Test::Unit::TestCase
     logger_mock.expects(:info).with('Appium driver initialized for:').once
     logger_mock.expects(:info).with('    project : local').once
     logger_mock.expects(:info).with(regexp_matches(/^\s{4}build\s{3}:\s\S{36}$/))
-    logger_mock.expects(:info).with(regexp_matches(/^\s{4}capabilities\s{4}:\s.+$/))
+    # logger_mock.expects(:info).with(regexp_matches(/^\s{4}capabilities\s{4}:\s.+$/))
     logger_mock
   end
 
@@ -218,7 +218,7 @@ class AppiumDriverTest < Test::Unit::TestCase
     driver.clear_and_send_keys_to_element('test_text_entry', 'Test_text')
   end
 
-  # TODO
+  # TODO Tests will follow in a subsequent PR after properly reassessing the logging
   # def test_environment_ids
   #   ENV['BRANCH_NAME'] = 'TEST BRANCH'
   #   ENV['BUILDKITE'] = 'true'