+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Enhancements
 
-- BrowserStack specific elements separated from AppAutomateDriver (now simply AppiumDriver)
+- BrowserStack specific elements separated from AppAutomateDriver (now simply AppiumDriver),
+  providing the ability to use MazeRunner with local devices.
   [#139](https://github.com/bugsnag/maze-runner/pull/139)
 
 # 2.7.0 - 2020/09/30